feat: new cmd add new cmd (get openmp include libs)
